feat: add basic telemetry
