+++ packages/create-bifrost-app/package.json
@@ -1,7 +1,7 @@
 {
   "name": "create-bifrost-app",
   "description": "Create an app based on the Bifrost template.",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "contributors": [],
   "license": "MIT",
   "homepage": "https://theodo-group.github.io/bifrost/",
@@ -23,7 +23,8 @@
   ],
   "scripts": {
     "build": "ncc build ./src/index.ts -o ./dist/ --minify --no-cache --no-source-map-register",
-    "dev": "ncc build ./src/index.ts -w -o dist/",
+    "dev": "ncc build ./src/indexNoTelemetry.ts -w -o dist/",
+    "dev:enable:telemetry": "ncc build ./src/index.ts -w -o dist/",
     "lint": "eslint",
     "lint:fix": "pnpm lint --fix",
     "test": "pnpm test:lint && pnpm test:type && pnpm test:unit && pnpm test:circular",
@@ -54,6 +55,7 @@
     "dependency-cruiser-config-custom": "workspace:*",
     "eslint-config-custom": "workspace:*",
     "got": "12.5.1",
+    "node-fetch": "3.3.1",
     "picocolors": "1.0.0",
     "prettier": "2.7.1",
     "rimraf": "3.0.2",
