+++ packages/create-bifrost-app/src/indexNoTelemetry.ts
@@ -1,0 +1,6 @@
+import { telemetryConfig } from 'helpers';
+import { main } from 'main';
+
+telemetryConfig.forceDisableTelemetry = true;
+
+main();
